requirements.txt
streamlit
requests
beautifulsoup4
langchain
sentence-transformers
faiss-cpu
python-dotenv